requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
django-cors-headers>=4.3,<5.0
